fix: sort data in groups query (#267)
https://github.com/uptrace/uptrace/issues/260
UI paginates the data by the order in which it was received in a response and then sorts it inside the page. This change sorts the data for proper pagination. It does not solve the issue of changing sort order inside the page which will be addressed on UI side.